--- src/main/java/cafeorder/dto/UpdateMemberDto.java
@@ -1,15 +1,0 @@
-package cafeorder.dto;
-
-import lombok.Getter;
-import org.springframework.web.bind.annotation.GetMapping;
-
-/**
- * author {yhh1056}
- * Create by {2020/10/14}
- */
-@Getter
-public class UpdateMemberDto {
-    private Long id;
-
-
-}
