+++ src/main/resources/static/css/createMemberForm.css
@@ -1,0 +1,68 @@
+.fieldError {
+    border-color: #bd2130;
+}
+
+body {
+    background-color: #f5f5f5;
+}
+
+.in-line {
+    width: 200px;
+    height: 40px;
+    margin: 0 auto;
+    text-align: center
+}
+
+.container {
+    text-align: center;
+}
+
+input {
+    display: inline-block;
+}
+
+input[type="text"] {
+    width: 100%;
+    height: 100%;
+    border: none;
+    font-size: 1em;
+    padding-left: 5px;
+    outline: none;
+    box-sizing: border-box;
+    color: black;
+}
+
+input[type=submit] {
+    width: 30%;
+    height: 100%;
+    margin-right:39%;
+    background-color: lightgray;
+    border: none;
+    background-color: white;
+    font-size: 1em;
+    color: #042AaC;
+    outline: none;
+    display: inline-block;
+    margin-left: -10px;
+    box-sizing: border-box;
+}
+input[type=button] {
+    width: 30%;
+    height: 100%;
+    margin-right: 0%;
+    background-color: lightgray;
+    border: none;
+    background-color: white;
+    font-size: 1em;
+    color: #042AaC;
+    outline: none;
+    display: inline-block;
+    margin-left: -10px;
+    box-sizing: border-box;
+
+}
+
+input[type=button]:hover {
+    background-color: lightgray;
+    display: inline-block;
+}
